+++ app/models/early_career_payments/eligibility_answers_presenter.rb
@@ -22,6 +22,9 @@ module EarlyCareerPayments
         a << nqt_in_academic_year_after_itt
         a << employed_as_supply_teacher
         a << has_entire_term_contract if eligibility.employed_as_supply_teacher?
+        a << employed_directly if eligibility.employed_as_supply_teacher?
+        a << subject_to_disciplinary_action
+        a << pgitt_or_ugitt_course
       end
     end
 
@@ -48,6 +51,30 @@ module EarlyCareerPayments
         translate("early_career_payments.questions.employed_as_supply_teacher"),
         (eligibility.employed_as_supply_teacher? ? "Yes" : "No"),
         "supply-teacher"
+      ]
+    end
+
+    def employed_directly
+      [
+        translate("early_career_payments.questions.employed_directly"),
+        (eligibility.employed_directly? ? "Yes" : "No"),
+        "employed-directly"
+      ]
+    end
+
+    def subject_to_disciplinary_action
+      [
+        translate("early_career_payments.questions.disciplinary_action"),
+        (eligibility.subject_to_disciplinary_action? ? "Yes" : "No"),
+        "disciplinary-action"
+      ]
+    end
+
+    def pgitt_or_ugitt_course
+      [
+        translate("early_career_payments.questions.postgraduate_itt_or_undergraduate_itt_course"),
+        translate("early_career_payments.answers.pgitt_or_ugitt_course.#{eligibility.pgitt_or_ugitt_course}"),
+        "postgraduate-itt-or-undergraduate-itt-course"
       ]
     end
   end