+++ db/migrate/20210421155453_add_pgitt_or_ugitt_course_to_early_career_payments_eligibilities.rb
@@ -1,0 +1,5 @@
+class AddPgittOrUgittCourseToEarlyCareerPaymentsEligibilities < ActiveRecord::Migration[6.0]
+  def change
+    add_column :early_career_payments_eligibilities, :pgitt_or_ugitt_course, :integer
+  end
+end