+++ db/migrate/20210422193758_add_employed_directly_to_early_career_payments_eligibilities.rb
@@ -1,0 +1,5 @@
+class AddEmployedDirectlyToEarlyCareerPaymentsEligibilities < ActiveRecord::Migration[6.0]
+  def change
+    add_column :early_career_payments_eligibilities, :employed_directly, :boolean
+  end
+end