+++ spec/factories/early_career_payments/eligibilities.rb
@@ -3,6 +3,8 @@ FactoryBot.define do
     trait :eligible do
       nqt_in_academic_year_after_itt { true }
       employed_as_supply_teacher { false }
+      subject_to_disciplinary_action { false }
+      pgitt_or_ugitt_course { :postgraduate }
     end
   end
 end