+++ spec/models/early_career_payments/eligibility_answers_presenter_spec.rb
@@ -4,7 +4,10 @@ RSpec.describe EarlyCareerPayments::EligibilityAnswersPresenter do
   let(:eligibility_attributes) do
     {
       nqt_in_academic_year_after_itt: true,
-      employed_as_supply_teacher: false
+      employed_as_supply_teacher: false,
+      employed_directly: true,
+      subject_to_disciplinary_action: false,
+      pgitt_or_ugitt_course: :postgraduate
     }
   end
   let(:eligibility) { claim.eligibility }
@@ -15,7 +18,9 @@ RSpec.describe EarlyCareerPayments::EligibilityAnswersPresenter do
   it "returns an array of questions and answers to be presented to the user for checking" do
     expected_answers = [
       [I18n.t("early_career_payments.questions.nqt_in_academic_year_after_itt"), "Yes", "nqt-in-academic-year-after-itt"],
-      [I18n.t("early_career_payments.questions.employed_as_supply_teacher"), "No", "supply-teacher"]
+      [I18n.t("early_career_payments.questions.employed_as_supply_teacher"), "No", "supply-teacher"],
+      [I18n.t("early_career_payments.questions.disciplinary_action"), "No", "disciplinary-action"],
+      [I18n.t("early_career_payments.questions.postgraduate_itt_or_undergraduate_itt_course"), "Postgraduate", "postgraduate-itt-or-undergraduate-itt-course"]
     ]
 
     expect(presenter.answers).to eq(expected_answers)
@@ -26,7 +31,10 @@ RSpec.describe EarlyCareerPayments::EligibilityAnswersPresenter do
       {
         nqt_in_academic_year_after_itt: true,
         employed_as_supply_teacher: true,
-        has_entire_term_contract: true
+        has_entire_term_contract: true,
+        employed_directly: true,
+        subject_to_disciplinary_action: false,
+        pgitt_or_ugitt_course: :postgraduate
       }
     end
 
@@ -34,7 +42,10 @@ RSpec.describe EarlyCareerPayments::EligibilityAnswersPresenter do
       expected_answers = [
         [I18n.t("early_career_payments.questions.nqt_in_academic_year_after_itt"), "Yes", "nqt-in-academic-year-after-itt"],
         [I18n.t("early_career_payments.questions.employed_as_supply_teacher"), "Yes", "supply-teacher"],
-        [I18n.t("early_career_payments.questions.has_entire_term_contract"), "Yes", "entire-term-contract"]
+        [I18n.t("early_career_payments.questions.has_entire_term_contract"), "Yes", "entire-term-contract"],
+        [I18n.t("early_career_payments.questions.employed_directly"), "Yes", "employed-directly"],
+        [I18n.t("early_career_payments.questions.disciplinary_action"), "No", "disciplinary-action"],
+        [I18n.t("early_career_payments.questions.postgraduate_itt_or_undergraduate_itt_course"), "Postgraduate", "postgraduate-itt-or-undergraduate-itt-course"]
       ]
 
       expect(presenter.answers).to eq(expected_answers)