+++ CHANGELOG.md
@@ -14,10 +14,13 @@ The format is based on [Keep a Changelog]
 - Build new Early Career Payments Journey with following page additions
   - Add Landing Page - 'Apply for an early-career payment'
   - Eligibility Journey
+    - Add 'Are you employed directly by your school?'
     - Add 'Did you do your NQT the academic year after your ITT?'
     - Add 'Are you currently employed as a Supply Teacher?'
     - Add 'Do you have a contract to teach at the same school for an entire term
       or longer?'
+    - Add 'Are you currently subject to disciplinary action?'
+    - Add 'Did you do a postgraduate ITT course or an undergraduate ITT course?'
     - Add 'You are not eligible for an early careers payment'
 - Automatically update admin claim tasks with DQT API